+++ src/app/features/components/login/EmailLogin.tsx
@@ -1,0 +1,71 @@
+'use client';
+
+import { useState } from 'react'
+import { Input } from "@/components/ui/input"
+import { Label } from "@/components/ui/label"
+import { Button } from "@/components/ui/button"
+import { Alert, AlertDescription } from "@/components/ui/alert"
+import { useRouter } from 'next/navigation'
+import { signinWithPassword } from '@/lib/supabase/supabaseFunctions'
+
+const EmailLogin = () => {
+
+    const [email, setEmail] = useState<string>('')
+    const [password, setPassword] = useState<string>('')
+    const [error, setError] = useState<string | null>(null)
+
+    const router = useRouter()
+
+    const handleSubmit = async (e: React.FormEvent) => {
+        e.preventDefault()
+        setError(null)
+        
+        if (!email || !password) {
+            setError('メールアドレスとパスワードを入力してください。')
+            return
+        }else{
+            const { error } = await signinWithPassword(email, password)
+            if(!error){
+                router.push('/main');
+            }
+        }
+    }
+
+    return (
+        <form onSubmit={handleSubmit}>
+            <div className="space-y-4">
+                <div className="space-y-2">
+                    <Label htmlFor="email">メールアドレス</Label>
+                    <Input
+                    id="email"
+                    type="email"
+                    placeholder="your@email.com"
+                    value={email}
+                    onChange={(e) => setEmail(e.target.value)}
+                    required
+                    />
+                </div>
+                <div className="space-y-2">
+                    <Label htmlFor="password">パスワード</Label>
+                    <Input
+                    id="password"
+                    type="password"
+                    value={password}
+                    onChange={(e) => setPassword(e.target.value)}
+                    required
+                    />
+                </div>
+                {error && (
+                    <Alert variant="destructive">
+                    <AlertDescription>{error}</AlertDescription>
+                    </Alert>
+                )}
+                <Button type="submit" className="w-full bg-black text-white">
+                    ログイン
+                </Button>
+            </div>
+        </form>
+    )
+}
+
+export default EmailLogin
